feat: useEffect 사용 x & 컴포넌트 분리
